Cleanup on tools/build if npm install fails.

+++ tools/build.sh
@@ -16,6 +16,11 @@ abort() {
   echo "$1" && exit 1
 }
 
+cleanup() {
+  cd $CURRENT_PATH
+  git checkout $current_branch
+}
+
 get_current_branch() {
   git branch --no-color 2> /dev/null | sed -e '/^[^*]/d' -e 's/* \(.*\)/\1/'
 }
@@ -61,6 +66,12 @@ build(){
   cd "$ROOT/$FOLDER"
 
   BUNDLE_ONLY=1 npm install --production # > /dev/null
+  if [ $? -ne 0 ]; then
+    echo "NPM install failed. Reverting changes..."
+    cleanup
+    return 1
+  fi
+
   rm -f npm-shrinkwrap.json
 
   # remove stuff from main tarball
@@ -92,8 +103,7 @@ build(){
   rm -fr "$ROOT"
   cd $DIST
   # ./checksum.sh $VERSION
-  cd $CURRENT_PATH
-  git checkout $current_branch
+  cleanup
 
 }
